requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
Django==4.2.6
djangorestframework==3.14.0
filelock==3.12.4
fsspec==2023.9.2
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
pydub==0.25.1
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
semantic-version==2.10.0
setuptools-rust==1.7.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.3
tomli==2.0.1
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
